# Remove dispatch optimization