@ry 2019-08-05
travis benchmarks for this branch:
```
{
  "req_per_sec": {
    "deno_tcp": 52181, 
    "deno_core_single": 54048, 
    "node_proxy_tcp": 20092, 
    "deno_proxy_tcp": 18167, 
    "deno_core_multi": 56638, 
    "deno_tcp_current_thread": 51249, 
    "hyper": 59218, 
    "node_tcp": 57930, 
    "deno_http": 19178, 
    "node_http": 24727, 
    "deno_proxy": 1663, 
    "node_proxy": 2740
  }, 
  "max_latency": {
    "deno_tcp": 8.82, 
    "deno_core_single": 0.496, 
    "node_proxy_tcp": 1.1, 
    "deno_proxy_tcp": 5.56, 
    "deno_core_multi": 1.3, 
    "deno_tcp_current_thread": 9.09, 
    "hyper": 0.846, 
    "node_tcp": 0.818, 
    "deno_http": 2.64, 
    "node_http": 1.96, 
    "deno_proxy": 30.85, 
    "node_proxy": 9.46
  }, 
```

For master branch
```
{
  "req_per_sec": {
    "deno_tcp": 51184, 
    "deno_core_single": 57055, 
    "node_proxy_tcp": 19587, 
    "deno_proxy_tcp": 17110, 
    "deno_core_multi": 51390, 
    "deno_tcp_current_thread": 48308, 
    "hyper": 61656, 
    "node_tcp": 56250, 
    "deno_http": 18386, 
    "node_http": 24297, 
    "deno_proxy": 1532, 
    "node_proxy": 2607
  }, 
  "max_latency": {
    "deno_tcp": 8.93, 
    "deno_core_single": 0.451, 
    "node_proxy_tcp": 1.2, 
    "deno_proxy_tcp": 6.0, 
    "deno_core_multi": 1.37, 
    "deno_tcp_current_thread": 9.57, 
    "hyper": 1.11, 
    "node_tcp": 0.684, 
    "deno_http": 2.8, 
    "node_http": 2.52, 
    "deno_proxy": 48.5, 
    "node_proxy": 10.81
  }, 
```